pyproject.toml
[project]
name = "rag-gist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.2.1",
    "langchain-pinecone>=0.2.13",
    "python-dotenv>=1.2.2",
]
